pyproject.toml
[project]
name = "burning-man-matchbot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    # Data layer
    "sqlmodel>=0.0.21",
    "alembic>=1.13.3",
    "aiosqlite>=0.17.0",
    "greenlet>=3.0.0",           # Required by SQLAlchemy async

    # Platform listeners
    "asyncpraw>=7.7.1",
    "discord.py>=2.4.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",

    # LLM clients
    "anthropic>=0.34.0",
    "openai>=1.47.0",

    # Config & validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.1",

    # CLI + terminal output
    "typer>=0.12.5",
    "rich>=13.8.0",

    # Scheduling + async utils
    "apscheduler>=3.10.4",
    "anyio>=4.6.0",

    # Templating
    "jinja2>=3.1.4",

    # YAML
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ipython>=9.10.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "respx>=0.21.0",
    "ruff>=0.15.1",
    "types-pyyaml>=6.0.12",
]

[project.scripts]
matchbot = "matchbot.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/matchbot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
